fix typescript issue with enums
