+++ frontend/tet/admin/src/backend-api/__tests__/transformations.test.ts
@@ -1,15 +1,12 @@
-import React from 'react';
+import { hdsDateToIsoFormat, isoDateToHdsFormat } from 'tet/admin/backend-api/transformations';
 
-// import { hdsDateToIsoFormat, isoDateToHdsFormat } from 'tet/admin/backend-api/transformations';
-//
 describe('transformations', () => {
-  it('works', () => {});
-  // it('can transform HDS date into ISO format', () => {
-  //   expect(hdsDateToIsoFormat('4.5.2022')).toBe('2022-05-04');
-  //   expect(hdsDateToIsoFormat('10.12.2022')).toBe('2022-12-10');
-  // });
-  // it('can transform ISO date into HDS format', () => {
-  //   expect(isoDateToHdsFormat('2022-12-10')).toBe('10.12.2022');
-  //   expect(isoDateToHdsFormat('2022-05-04')).toBe('4.5.2022');
-  // });
+  it('can transform HDS date into ISO format', () => {
+    expect(hdsDateToIsoFormat('4.5.2022')).toBe('2022-05-04');
+    expect(hdsDateToIsoFormat('10.12.2022')).toBe('2022-12-10');
+  });
+  it('can transform ISO date into HDS format', () => {
+    expect(isoDateToHdsFormat('2022-12-10')).toBe('10.12.2022');
+    expect(isoDateToHdsFormat('2022-05-04')).toBe('4.5.2022');
+  });
 });
