+++ frontend/tet/admin/src/hooks/backend/useKeywordType.ts
@@ -1,4 +1,4 @@
-import { ClassificationType, KeywordFn } from 'tet/admin/types/classification';
+import { ClassificationType, KeywordFn } from 'tet/admin/types/keywords';
 import { useQueries } from 'react-query';
 import { getWorkFeatures, getWorkMethods } from 'tet/admin/backend-api/linked-events-api';
