+++ frontend/tet/admin/src/types/classification.d.ts
@@ -4,10 +4,3 @@ export type OptionType = Option & {
   label: string;
   value: string;
 };
-
-export enum ClassificationType {
-  WORKING_METHOD = 'WORKING_METHOD',
-  WORKING_FEATURE = 'WORKING_FEATURE',
-  KEYWORD = 'KEYWORD',
-}
-export type KeywordFn = (keywordUrl: string) => ClassificationType;
