+++ frontend/tet/admin/src/types/keywords.ts
@@ -1,0 +1,8 @@
+// Note: `export enum` inside a `.d.ts` file results in module not found error
+export enum ClassificationType {
+  WORKING_METHOD = 'WORKING_METHOD',
+  WORKING_FEATURE = 'WORKING_FEATURE',
+  KEYWORD = 'KEYWORD',
+}
+
+export type KeywordFn = (keywordUrl: string) => ClassificationType;
